+++ src/components/genres-list/genres-list.jsx
@@ -3,11 +3,11 @@ import PropTypes from 'prop-types';
 
 import proptypes from '../../type';
 import {DEFAULT_MOVIES_FILTER_VALUE} from '../../const';
-import {ActionCreator} from '../../store/actions';
+import {changeGenre, filterMoviesByGenre} from '../../store/action';
 import {connect} from 'react-redux';
 
 const GenresList = (props) => {
-  const {currentGenre, genres, changeGenre} = props;
+  const {currentGenre, genres, changeGenreAction} = props;
 
   return (
     <ul className="catalog__genres-list">
@@ -17,7 +17,7 @@ const GenresList = (props) => {
             <a
               onClick={(evt) => {
                 evt.preventDefault();
-                changeGenre(genre);
+                changeGenreAction(genre);
               }}
               href="#"
               className="catalog__genres-link">
@@ -33,7 +33,7 @@ const GenresList = (props) => {
 GenresList.propTypes = {
   genres: proptypes.genres,
   currentGenre: proptypes.genre,
-  changeGenre: PropTypes.func.isRequired
+  changeGenreAction: PropTypes.func.isRequired
 };
 
 const mapStateToProps = (state) => {
@@ -49,9 +49,9 @@ const mapStateToProps = (state) => {
 };
 
 const mapDispatchToProps = (dispatch) => ({
-  changeGenre(genre) {
-    dispatch(ActionCreator.changeGenre(genre));
-    dispatch(ActionCreator.filterMoviesByGenre());
+  changeGenreAction(genre) {
+    dispatch(changeGenre(genre));
+    dispatch(filterMoviesByGenre());
   },
 
 });
