--- src/store/actions.js
@@ -1,18 +1,0 @@
-export const ActionType = {
-  CHANGE_GENRE: `CHANGE_GENRE`,
-  FILTER_MOVIES_BY_GENRE: `FILTER_MOVIES_BY_GENRE`
-};
-
-export const ActionCreator = {
-  changeGenre: (genre) => {
-    return {
-      type: ActionType.CHANGE_GENRE,
-      payload: genre
-    };
-  },
-  filterMoviesByGenre: () => {
-    return {
-      type: ActionType.FILTER_MOVIES_BY_GENRE
-    };
-  }
-};
